requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile
#
asgiref==3.3.1            # via django
django-polymorphic==3.0.0  # via -r requirements.in
django==3.1.3             # via -r requirements.in, django-polymorphic
pillow==8.0.1             # via -r requirements.in
pytz==2020.4              # via django
sqlparse==0.4.1           # via django
